Рефакторинг utils. Разбиение на несколько модулей вспомогательных функций.
